requirements.txt
torch==1.11.0
transformers==4.15.0
timm==0.4.12
scipy==1.4.1
fairscale==0.4.4
pycocoevalcap
ruamel_yaml
yacs 
ujson 
einops 
fnvhash 
pillow 
tqdm 
torchmetrics 
albumentations
opencv-python-headless==4.5.5.64
pytorch-lightning==1.5.4
